DAG recursion tree traversal replaces older verification method
- Checkpoint method implemented:
When DAG recursion tree is traversed, for every root node a new image is created.
The subtree of that node is tried on the new image and in turn creates other images
for sub sub trees and so on. Once the whole subtree is tried, the node docker image
is discarded.

- Only a single actor system is intantiated on verifier on host machine
- shutdown message added to application running in docker container to halt its actor system

DAG recursion tree traversal replaces older verification method

Checkpoint method implemented:
When DAG recursion tree is traversed, for every root node a new image is created.
The subtree of that node is tried on the new image and in turn creates other images
for sub sub trees and so on. Once the whole subtree is tried, the node docker image
is discarded.

recur tree cleanup

handle shutdown message and shutdown actor system upon receipt