new throttling design checkpoint: Compiling, havnt tested yet
The earlier concurrent design using futures did not behave well
with resource allocation as the concurrency was unbounded. As the
permutation tree was traversed the number of simultaneous docker
containers exploded. This resulted in resource allocation failures
like response timeouts, disk space issues etc.

The new design is concurrent but the amount of concurrency is throttled
by means of tokens (Token bucket rate control mechanism). Akka is used
for concurrent execution of resources.
A high level overview of design is :-

- There are fixed number of worker actors

- Each worker actor is assigned a resource to install and report back result

- A worker is a combination of two actors: one manages container lifecycle and
the other coordinates with remote actor system on docker container and drives the
container fsm actor

- There is a throttling actor that throttles producer by means of distributing
tokens

- Its a pull based architecture, whenever a worker actor becomes free to take up a
new resource, it indicates it to the throttler. The throttler then aggregates such
requests and passes corresponding tokens to producer

- producer buffers some amount of work and as soon as tokens are made available,
dispatches the bufferred work to workers.