Fix for thias-bind-buggy
The domain of paths is calculated from the read and write set. Earlier,
the read set of TestFileState(p, _) would be all ancestors of p.
Since this is no longer the case, it is now possible for the parent of
a path to not be in the domain of paths. E.g., for mkdir(/a/b/c),
if nothing else reads or writes to /a, then /a will not be in the domain.

This patch accounts for this case and also eliminates assertions of
the form X => dir("/"), since dir("/") is already asserted.